fix: resolve all CI compilation errors
- MainActivity.kt: replace specific icon imports with wildcard (Icons.Filled.*, Icons.Outlined.*)
- InnerTubeFeed.kt: cast StreamInfoItem.duration Long to Int for VideoItem
- StreamResolver.kt: fix NewPipeExtractor API - Description.content (not textContent), VideoStream.isVideoOnly (not videoOnly)
- PiTubeSplashScreen.kt: add missing Modifier.scale import

+++ app/src/main/java/com/omersusin/pitube/data/InnerTubeFeed.kt
@@ -33,7 +33,7 @@ object InnerTubeFeed {
                         uploaderAvatar = "",
                         uploaderUrl = item.uploaderUrl ?: "",
                         channelId = "",
-                        duration = item.duration,
+                        duration = item.duration.toInt(),
                         views = item.viewCount,
                         uploadedDate = item.textualUploadDate ?: "",
                         isShort = false
